Layout_TestCase/http/tests/xmlhttprequest/origin-whitelisting-removal-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
Tests the behavior of whitelisting origins and removing them later.

Testing: source origin: http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests destination origin: http:localhost 
Loading: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101
Loading: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101


Testing: source origin: http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests destination origin: http:localhost allowing subdomains
Loading: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101
Loading: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101


Testing: source origin: http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests destination origin: hTtP:LoCaLhOsT allowing subdomains
Loading: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101
Loading: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101


Testing: source origin: http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests destination origin: http: allowing subdomains
Loading: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101
Loading: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101
